added new actionbar to feed
